Add gender field to user and extend functionality
Added a Gender property to the User object to provide more detailed user profiling. This required adding a new migration which adds the Gender column to the Users table in the database. Files related to this operation include user's C# model, migration files, and database snapshot. Other changes include expansion of the functionality of the application with the addition of new About page, User Controller and Swagger for better API documentation and testing. Some namespaces and dependencies have been updated accordingly. A LoginDTO class was also added under DataTransfer to handle authentication processes.