Remove all migration files, add UserData migration
All existing migration files were removed as there were changes in the model that couldn't be migrated. The UserData migration file was added to facilitate the users' password encryption. The migration involves creation of UserRoles and Roles tables and population of defined roles in the Roles table. The change is aimed at improving the users' data security by storing only encrypted passwords.

The RegistrationDTO class was also added to facilitate registration process and it includes Email, UserName, Password, Password2, and DateOfBirth fields.