"Updated user registration and authentication flow"
This commit implements updates to the user registration and authentication flow in our EpochApp. The DateOfBirth field in RegistrationDTO is now optional to account for possible null values. The JWT token is now read more accurately in EpochUserService with the claim type corrected to be "jwt". The user model now includes NormalizedUserName and NormalizedEmailName for more effective case-insensitive management.

In the client-side, complete registration form in Registration.razor has been added for user sign-up along with the necessary form field bindings and form validation. LogOut.razor now includes the actual logout functionality with navigation after logout. In addition, Log information has been included in EpochAuthProvider to keep track of user signIn and signOut.

ClaimTypes.NameIdentifier is added to the user's claims list in UserData to store the UserID for secure access and retrieval of user data.