"Refactor shared files structure to improve readability
The shared file structure 'World.cs' was removed and a full refactor of the 'EpochApp/Shared' directory was done for better organization. This includes the creation of new classes like 'WorldDate.cs', 'WorldMeta.cs', and new enums like 'ContainerWidthType.cs'. The 'World.cs' was improved with additional features. 'Utils.cs' was updated to be a static class and the file locations of multiple 'Blog' classes and 'Lookups' were moved to better match the application structure. Lastly, the 'NormalizedEmailName' property of the 'User.cs' class was renamed to 'NormalizedEmail' for better clarity and understanding."