Remove BlogOwners migration
The migration "BlogOwners" was removed as it was determined to be unnecessary and not aligning with the desired database design. This will simplify the DB structure and improve the performance. Future database interactions will not have to consider the BlogOwners table, reducing complexity and potential errors. The related Designer file was also removed accordingly.