Update SDK version and integrate Flatify project
Updated the SDK version in global.json from 7.0.0 to 8.0.0 for our development environment. We included a reference to the Flatify project and its ThemeProvider in MainLayout. New ContainerBase component was also added under EpochComponents. Minor improvements were made to the naming in the LoginKit and a new SVG icon was added to the .idea directory.