Add new Accordion component to Flatify
The update involves adding a new Accordion component to the Flatify library. This new component, which is designed to encapsulate multiple sections of content, consists of several parts: FlatAccordion.razor, AccordionItem.razor, AccordionItem.razor.cs and FlatAccordion.razor.cs. Also updated is the FlatifyComponentTest page to include usage of the new component.