Add user Worlds page and update navigation and forms
A new page for displaying user Worlds has been created and a link to this page has been added to the navigation menu. Various changes were also made to improve readability and functionality of the registration and login forms. In addition to this, an API method to fetch user's Worlds was added to the WorldsController.