Refactor About page to replace FlatCard with FlatTypo
The About page has undergone a simplification by replacing the previous "FlatCard" components with a "FlatTypo" component. This aims to simplify the presentational aspect of our site and provide better readability of the About page contents. The changes also include an addition of a "FlatBreak" line with accent color to enhance visual separation.