Implement mail service and update controller routes
This update adds a new mail service to the project which can be used for mailing tasks such as sending notifications or recovering passwords. The service is registered in the services container for dependency injection. Additionally, modifications were made to the routes in ContactController and BlogsController to adjust parameter types and add new functionalities. A MailController has also been added to manage mail related requests.