Refactor navigation and implement feature to delete worlds
The navigation has been refactoren in Razor pages using MudBlazor toolbar for a more intuitive design, and a feature to delete created worlds has been implemented. This involved updates in multiple Razor components, the addition of a handler for the delete action in the WorldsController, and modifications to other server-side and client-side logic to accommodate for this new feature.