Add authorization checks, new blog view and policy page
Added authorization checks to the navigation bar and created a new `BlogView` component for displaying blog posts. Also introduced a `Policy` page with embedded PDFs for privacy policy. Some minor UI enhancements were also made, including changes to the footer and the app bar, and adding new directories and files to the project.