Add new classes and update existing ones for world creation functionality
This commit includes the addition of new DTOs and Services for world creation. It also adds a new lookup service to fetch data from lookup tables. Coding was done to support the addition of new files, languages, and form submission for creating a new world. Some existing classes were modified to include more fields for handling world genres and files. Changes were also made in the LookupController to fetch the required data.