Update design and interaction in WorldSelector component
The WorldSelector component has been significantly updated. The layout of the component has been redesigned with a more interactive and intuitive UI which now changes based on whether the navigation is collapsed or not. In addition, the ShouldRender method was overridden in the WorldSelector code-behind to optimize rendering when the 'Condensed' state changes.