Update articles view and integrate World refinements
This commit includes substantial updates to the articles view including properly mapping World to UserWorld. A new table of contents feature for articles has been added. ActiveWorld is now directly fetched for user due to the removal of OnWorldChanged event. Additionally, updates have been made to ensure timestamps are in UTC and a new Article endpoint for a specific world and article combination was introduced.