Add extensive article functionality and enhancements
The code significantly expands various functionalities related to articles. A notable change is the creation of the ArticleService that handles the mechanics of fetching, displaying, and creating article related data, including details about the author, its sections, and related tags. Furthermore, improvements are made to the handling of tags with the TagService, which now has increased methods for creating and managing different kinds of tags. Lastly, the ArticlesController has been simplified and now utilizes the aforementioned services for its operations.