Refactor and simplify blog-related classes
This commit includes refactoring and simplifying of blog-related classes in the shared and server projects. The modifications involve deletion of unused files, simplifying certain properties, and renaming files to follow a consistent naming convention. A new class, PostComment, has also been added to cater to the new requirements in the Blog's feature set. Necessary changes in the configurations of the DbContext to accommodate these changes have also been implemented.