"Implement manuscript service and update UI elements"
In this update, we added a manuscript service for managing user's manuscripts. This entails getting all manuscripts for a specific user, creating, updating, and retrieving a specific manuscript. Moreover, there were significant user interface changes as well, including modification to the navbar, subnav, MainLayout, and DashLayout. Added new icons to the StaticUtils as AwesomeIcons along with a manuscripts page for Dashboard in the client. Furthermore, we modified ArticleService to include calls to the new manuscript service, and modified the nav references to reflect the new page additions.